backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
channels>=4.0
psycopg2>=2.9
pyzbar>=0.1.8
pytesseract>=0.3.9
opencv-python>=4.6.0.66
numpy>=1.23
scikit-learn>=1.2
